My Contribution for hacktoberfest 2022
Contribution from @ashutosh7i for hacktoberfest 2022
