+++ C++/Rain_Building.cpp
@@ -1,0 +1,22 @@
+#include <iostream>
+#include <vector>
+
+using namespace std;
+
+void Rain_Building()
+{
+
+}
+
+
+
+
+
+
+
+
+
+int main()
+{
+
+}
