--- Neon.html
@@ -1,22 +1,0 @@
-<!-- ARTICLE URL = https://css-tricks.com/how-to-create-neon-text-with-css/ -->
-
-<!DOCTYPE html>
-<html>
-
-<head>
-    <meta charset="utf-8" />
-    <title>
-        VG_WebPage
-    </title>
-
-    <link rel="stylesheet" type="text/css" href="neon.css">
-</head>
-
-<div class="container">
-    <h1 class="neonText">
-          neon <br> lights
-    </h1>
-
- </div>
-
-</html>
