--- Rain_Building.cpp
@@ -1,22 +1,0 @@
-#include <iostream>
-#include <vector>
-
-using namespace std;
-
-void Rain_Building()
-{
-
-}
-
-
-
-
-
-
-
-
-
-int main()
-{
-
-}
