--- neon.css
@@ -1,442 +1,0 @@
-/*.neonText {
-    color: #fff;
-    text-shadow:
-        0 0 7px #fff,
-        0 0 10px #fff,
-        0 0 21px #fff,
-        0 0 42px #0fa,
-        0 0 82px #0fa,
-        0 0 92px #0fa,
-        0 0 102px #0fa,
-        0 0 151px #0fa;
-  }
-  
-//ADDITIONAL STYLING
-    
-  body {
-    font-size: 18px;
-    font-family: "Vibur", sans-serif;
-    background-color: #010a01;
-  }  
-  
-  h1, h2 {
-    text-align: center;
-    text-transform: uppercase;
-    font-weight: 400;
-  }
-    
-  h1 {
-      font-size: 4.2rem;
-  }
-    
-  h2 {
-      font-size: 1.8rem;
-  }
-  
-  .container {
-    margin-top: 20vh;
-  }
-  */
-  
-/*FLIKRING EFFECT =========================================================================*/
-/* .neonText {
-    animation: flicker 1.5s infinite alternate;
-    color: #fff;
-  }
-  
-  h1 {
-    font-size: 4.2rem;    
-  }
-  
-  // Flickering animation 
-  @keyframes flicker {
-      
-    0%, 18%, 22%, 25%, 53%, 57%, 100% {
-  
-        text-shadow:
-        0 0 4px #fff,
-        0 0 11px #fff,
-        0 0 19px #fff,
-        0 0 40px #0fa,
-        0 0 80px #0fa,
-        0 0 90px #0fa,
-        0 0 100px #0fa,
-        0 0 150px #0fa;
-    
-    }
-    
-    20%, 24%, 55% {        
-        text-shadow: none;
-    }    
-  }
-  
-  
-// Additional styling 
-    
-  h2 {
-   font-size: 1.8rem;
-  }
-  .container {
-    margin-top: 20vh;
-  }
-  
-  body {
-    font-size: 18px;
-    font-family: "Helvetica Neue", sans-serif;
-    background-color: #010a01;
-  }  
-  
-  h1, h2 {
-    text-align: center;
-    text-transform: uppercase;
-    font-weight: 400;
-  } */
-
-  /* subtle flicker ===================================================================*/
-  /* .neonText {
-    color: #fff;
-    text-shadow:
-        0 0 7px #fff,
-        0 0 10px #fff,
-        0 0 21px #fff,
-        0 0 42px #0fa,
-        0 0 82px #0fa,
-        0 0 92px #0fa,
-        0 0 102px #0fa,
-        0 0 151px #0fa;
-  }
-  
-  h1 {
-    font-size: 4.2rem;
-    animation: flicker 1.5s infinite alternate;     
-  }
-  
-  //Flickering animation
-  @keyframes flicker {
-      
-    0%, 18%, 22%, 25%, 53%, 57%, 100% {
-  
-        text-shadow:
-        0 0 4px #fff,
-        0 0 11px #fff,
-        0 0 19px #fff,
-        0 0 40px #0fa,
-        0 0 80px #0fa,
-        0 0 90px #0fa,
-        0 0 100px #0fa,
-        0 0 150px #0fa;
-    
-    }
-    
-    20%, 24%, 55% {        
-        text-shadow: none;
-    }    
-  }
-  
-  
-  //Additional styling
-    
-  h2 {
-   font-size: 1.8rem;
-  }
-  .container {
-    margin-top: 20vh;
-  }
-  
-  body {
-    font-size: 18px;
-    font-family: "Helvetica Neue", sans-serif;
-    background-color: #010a01;
-  }  
-  
-  h1, h2 {
-    text-align: center;
-    text-transform: uppercase;
-    font-weight: 400;
-  } */
-  
-  /* Pulsating glow =================================================*/
-  /* .neonText {
-    color: #fff;
-    text-shadow:
-      0 0 7px #fff,
-      0 0 10px #fff,
-      0 0 21px #fff,
-      0 0 42px #0fa,
-      0 0 82px #0fa,
-      0 0 92px #0fa,
-      0 0 102px #0fa,
-      0 0 151px #0fa;
-  }
-
-
-//Additional styling
-  
-body {
-  font-size: 18px;
-  font-family: "Helvetica Neue", sans-serif;
-  background-color: #010a01;
-}  
-
-h1, h2 {
-  text-align: center;
-  text-transform: uppercase;
-  font-weight: 400;
-}
-  
-h1 {
-    font-size: 4.2rem;
-}
-
-.pulsate {
-     animation: pulsate 1.5s infinite alternate;     
-}
-  
-h2 {
-    font-size: 1.8rem;
-}
-
-.container {
-  margin-top: 20vh;
-}
-
-@keyframes pulsate {
-    
-  100% {
-
-      text-shadow:
-      0 0 4px #fff,
-      0 0 11px #fff,
-      0 0 19px #fff,
-      0 0 40px #0fa,
-      0 0 80px #0fa,
-      0 0 90px #0fa,
-      0 0 100px #0fa,
-      0 0 150px #0fa;
-  
-  }
-  
-  0% {
-
-    text-shadow:
-    0 0 2px #fff,
-    0 0 4px #fff,
-    0 0 6px #fff,
-    0 0 10px #0fa,
-    0 0 45px #0fa,
-    0 0 55px #0fa,
-    0 0 70px #0fa,
-    0 0 80px #0fa;
-
-} */
-
-/* Subtle flicker ========================================================*/
-
-/* .neonText {
-    color: #fff;
-    text-shadow:
-      0 0 7px #fff,
-      0 0 10px #fff,
-      0 0 21px #fff,
-      0 0 42px #f09,
-      0 0 82px #f09,
-      0 0 92px #f09,
-      0 0 102px #f09,
-      0 0 151px #f09;
-  }
-
-
-//Additional styling
-  
-body {
-  font-size: 18px;
-  font-family: "Vibur", sans-serif;
-  background-color: #010a01;
-}  
-
-h1, h2 {
-  text-align: center;
-  font-weight: 400;
-  line-height: 1;
-}
-  
-h1 {
-    font-size: 6.2rem;
-    animation: pulsate 0.11s ease-in-out infinite alternate;     
-}
-  
-h2 {
-    font-size: 1.8rem;
-}
-
-.container {
-  margin-top: 20vh;
-}
-
-@keyframes pulsate {
-    
-  100% {
-
-      text-shadow:
-      0 0 4px #fff,
-      0 0 11px #fff,
-      0 0 19px #fff,
-      0 0 40px #f09,
-      0 0 80px #f09,
-      0 0 90px #f09,
-      0 0 100px #f09,
-      0 0 150px #f09;
-  
-  }
-  
-  0% {
-
-    text-shadow:
-    0 0 4px #fff,
-    0 0 10px #fff,
-    0 0 18px #fff,
-    0 0 38px #f09,
-    0 0 73px #f09,
-    0 0 80px #f09,
-    0 0 94px #f09,
-    0 0 140px #f09;
-
-}
-} */
-
-/* Using a background image ====================================================== */
-/* .neonText {
-    color: #fff;
-    text-shadow:
-    0 0 7px #fff,
-    0 0 10px #fff,
-    0 0 21px #fff,
-    0 0 42px #f09,
-    0 0 82px #f09,
-    0 0 92px #f09,
-    0 0 102px #f09,
-    0 0 151px #f09;
-  }
-    
-  body {
-    font-size: 18px;
-    font-family: "Yellowtail", sans-serif;
-    background-color: #010a01;
-    background-image: url("https://cdn.pixabay.com/photo/2017/02/12/12/42/wall-2059909_640.png")
-  }  
-  
-  h1, h2 {
-    text-align: center;
-    font-weight: 400;
-  }
-    
-  h1 {
-      font-size: 5.2rem;
-      animation: pulsate 0.11s ease-in-out infinite alternate;     
-  }
-    
-  h2 {
-      font-size: 1.8rem;
-  }
-  
-  @keyframes pulsate {
-      
-    100% {
-  
-        text-shadow:
-        0 0 7px #fff,
-        0 0 10px #fff,
-        0 0 21px #fff,
-        0 0 42px #f09,
-        0 0 82px #f09,
-        0 0 92px #f09,
-        0 0 102px #f09,
-        0 0 151px #f09;
-    
-    }
-    
-    0% {
-  
-      text-shadow:
-      0 0 7px #fff,
-      0 0 10px #fff,
-      0 0 21px #fff,
-      0 0 42px #f09,
-      0 0 82px #f09,
-      0 0 92px #f09,
-      0 0 102px #f09,
-      0 0 151px #f09;
-  
-  } */
-
-  /* Adding a border ============================================= */
-  .neonText {
-    color: #fff;
-    text-shadow:
-      0 0 7px #fff,
-      0 0 10px #fff,
-      0 0 21px #fff,
-      0 0 42px #bc13fe,
-      0 0 82px #bc13fe,
-      0 0 92px #bc13fe,
-      0 0 102px #bc13fe,
-      0 0 151px #bc13fe;
-  }
-    
-  body {
-    font-size: 18px;
-    font-family: "Sacramento", sans-serif;
-    background-color: #010a01;
-    display: flex;
-    justify-content: center;
-    align-items: center;  
-  }  
-  
-  h1, h2 {
-    text-align: center;
-    font-weight: 400;
-  }
-    
-  h1 {
-    font-size: 6.2rem;
-    animation: pulsate 1.5s infinite alternate;  
-    border: 0.2rem solid #fff;
-    border-radius: 2rem;
-    padding: 0.4em;
-    box-shadow: 0 0 .2rem #fff,
-              0 0 .2rem #fff,
-              0 0 2rem #bc13fe,
-              0 0 0.8rem #bc13fe,
-              0 0 2.8rem #bc13fe,
-              inset 0 0 1.3rem #bc13fe; 
-  }
-  
-  @keyframes pulsate {
-      
-    100% {
-  
-        text-shadow:
-        0 0 4px #fff,
-        0 0 11px #fff,
-        0 0 19px #fff,
-        0 0 40px #bc13fe,
-        0 0 80px #bc13fe,
-        0 0 90px #bc13fe,
-        0 0 100px #bc13fe,
-        0 0 150px #bc13fe;
-    
-    }
-    
-    0% {
-  
-      text-shadow:
-      0 0 2px #fff,
-      0 0 4px #fff,
-      0 0 6px #fff,
-      0 0 10px #bc13fe,
-      0 0 45px #bc13fe,
-      0 0 55px #bc13fe,
-      0 0 70px #bc13fe,
-      0 0 80px #bc13fe;
-  
-  }
